Added order percentage support to paymentprovider, using new Percentage field at PaymentProviderMethod.

+++ Core/uWebshop.Domain/Model/OrderDTO/Order.cs
@@ -44,7 +44,7 @@ namespace uWebshop.Domain.OrderDTO
 	    public DateTime? FulfillDate;
 		public PaymentInfo PaymentInfo;
 		public int PaymentProviderPrice;
-		public int? PaymentProviderOrderPercentage;
+		public decimal? PaymentProviderOrderPercentage;
 		public ValidateSaveAction ReValidateSaveAction;
 		public int RegionalVatAmount;
 		public bool? RevalidateOrderOnLoad;
